refactor: Use testFixtures instead of separate test-utils module
