+++ orcid-commons/src/main/java/no/sikt/nva/orcid/commons/OrcidConstants.java
@@ -1,6 +1,8 @@
-package no.sikt.nva.orcid.constants;
+package no.sikt.nva.orcid.commons;
 
 public final class OrcidConstants {
 
     public static final String ORCID_PRIMARY_PARTITION_KEY = "orcid";
+
+    private OrcidConstants() {}
 }
