+++ orcid-commons/src/main/java/no/sikt/nva/orcid/commons/service/ReadOrcidCredentialsService.java
@@ -1,7 +1,7 @@
 package no.sikt.nva.orcid.commons.service;
 
 import static java.util.Objects.isNull;
-import static no.sikt.nva.orcid.constants.OrcidConstants.ORCID_PRIMARY_PARTITION_KEY;
+import static no.sikt.nva.orcid.commons.OrcidConstants.ORCID_PRIMARY_PARTITION_KEY;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.amazonaws.services.dynamodbv2.model.GetItemRequest;
