+++ orcid-commons/src/main/java/no/sikt/nva/orcid/commons/service/ServiceWithTransactions.java
@@ -1,6 +1,6 @@
 package no.sikt.nva.orcid.commons.service;
 
-import static no.sikt.nva.orcid.constants.OrcidConstants.ORCID_PRIMARY_PARTITION_KEY;
+import static no.sikt.nva.orcid.commons.OrcidConstants.ORCID_PRIMARY_PARTITION_KEY;
 import static nva.commons.core.attempt.Try.attempt;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.model.Put;
