--- orcid-testutils/build.gradle
@@ -1,13 +1,0 @@
-plugins {
-    id 'nva.orcid.client.java-conventions'
-}
-
-dependencies {
-    api(nvaCatalog.aws.sdk.dynamo)
-    api(nvaCatalog.junit.jupiter.api)
-    implementation(nvaCatalog.aws.dynamodb.local)
-    implementation(nvaLibs.core)
-    implementation(project(":orcid-constants"))
-    testRuntimeOnly(nvaCatalog.junit.jupiter.engine)
-    testRuntimeOnly(nvaCatalog.junit.platform.launcher)
-}
