+++ settings.gradle
@@ -38,5 +38,3 @@ dependencyResolutionManagement {
 rootProject.name = 'nva-orcid-client'
 include 'orcid-rest'
 include 'orcid-commons'
-include 'orcid-testutils'
-include 'orcid-constants'
